pyproject.toml
[tool.poetry]
name = "fastapi-translation-service"
version = "0.1.0"
description = ""
authors = ["Arystan Igen <arystan.igen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
motor = "^3.5.1"
httpx = "^0.27.0"
odmantic = "^1.0.2"
environ-config = "^23.2.0"
ipython = "^8.26.0"
gunicorn = "^22.0.0"
respx = "^0.21.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
pytest = "^8.2.2"
mypy-protobuf = "^3.6.0"
safety = "^3.2.4"
bandit = "^1.7.9"
ruff = "^0.5.1"
pytest-asyncio = "^0.23.7"
pytest-cov = "^4.1.0"
pytest-sugar = "^0.9.7"
pytest-xdist = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
